rpbnb: Random-Parameter Bivariate Negative Binomial Regression

Maximum-likelihood estimation of bivariate negative binomial models with Famoye/Sarmanov or discrete-copula (Frank, Gaussian, Clayton) dependence (see fit_bnb(), copula()), and maximum-simulated-likelihood estimation of a bivariate random-parameter negative binomial model under either dependence structure.

Details

Two estimation engines are available for the random-parameter model: fit_rpbnb() (Rcpp/OpenMP simulated likelihood, maxLik BFGS, supports offset()) and fit_rpbnb_tmb() (TMB automatic differentiation, nlminb, adds a Laplace approximation and dependence profiling). rpbnb() is a common front end that dispatches to either.

Author

Maintainer: Zhenyu Wang wonstran@hotmail.com

Authors:

Zhenyu Wang wonstran@hotmail.com